ref: updated predictions renderer
